+++ packages/components/src/org.standardnotes.advanced-checklist/src/app/hooks.ts
@@ -5,10 +5,7 @@ import type { RootState, AppDispatch } from './store'
 export const useAppDispatch = () => useDispatch<AppDispatch>()
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector
 
-export const useDidMount = (
-  effect: React.EffectCallback,
-  deps?: React.DependencyList
-) => {
+export const useDidMount = (effect: React.EffectCallback, deps?: React.DependencyList) => {
   const [didMount, setDidMount] = useState(false)
 
   useEffect(() => {