+++ packages/components/src/org.standardnotes.advanced-checklist/src/common/components/icons/DottedCircleIcon.tsx
@@ -6,14 +6,7 @@ export const DottedCircleIcon = () => {
       viewBox="0 0 20 20"
       xmlns="http://www.w3.org/2000/svg"
     >
-      <rect
-        x="0.5"
-        y="0.5"
-        width="19"
-        height="19"
-        rx="9.5"
-        strokeDasharray="2 2"
-      />
+      <rect x="0.5" y="0.5" width="19" height="19" rx="9.5" strokeDasharray="2 2" />
     </svg>
   )
 }