+++ packages/components/src/org.standardnotes.advanced-checklist/src/features/tasks/TaskItemList.test.tsx
@@ -52,9 +52,7 @@ it('renders the completed tasks container', () => {
 
   testRender(<TaskItemList group={groupWithCompletedTask} />)
 
-  const completedTasksContainer = screen.getByTestId(
-    'completed-tasks-container'
-  )
+  const completedTasksContainer = screen.getByTestId('completed-tasks-container')
 
   expect(completedTasksContainer).toBeInTheDocument()
   expect(completedTasksContainer).toHaveTextContent('completed tasks')