+++ packages/components/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/standardnotes/app/compare/@standardnotes/components-meta@0.1.4...@standardnotes/components-meta@0.2.0) (2022-06-11)
+
+
+### Features
+
+* add themes source ([#1089](https://github.com/standardnotes/app/issues/1089)) ([e80eb44](https://github.com/standardnotes/app/commit/e80eb441a717b7c892307635ff3af4db2e25d57c))
+
+
+
+
+
+## [0.1.4](https://github.com/standardnotes/app/compare/@standardnotes/components-meta@0.1.3...@standardnotes/components-meta@0.1.4) (2022-06-11)
+
+**Note:** Version bump only for package @standardnotes/components-meta
+
+
+
+
+
+## [0.1.3](https://github.com/standardnotes/app/compare/@standardnotes/components-meta@0.1.2...@standardnotes/components-meta@0.1.3) (2022-06-11)
+
+**Note:** Version bump only for package @standardnotes/components-meta
+
+
+
+
+
+## [0.1.2](https://github.com/standardnotes/app/compare/@standardnotes/components-meta@0.1.1...@standardnotes/components-meta@0.1.2) (2022-06-11)
+
+**Note:** Version bump only for package @standardnotes/components-meta
+
+
+
+
+
 ## [0.1.1](https://github.com/standardnotes/app/compare/@standardnotes/components-meta@0.1.0...@standardnotes/components-meta@0.1.1) (2022-06-10)
 
 **Note:** Version bump only for package @standardnotes/components-meta